// app/server/applicationsDB.server/applications/azahar/types.ts
export type AzaharButtonId =
  | "button_up"
  | "button_left"
  | "button_down"
  | "button_right"
  | "button_a"
  | "button_b"
  | "button_x"
  | "button_y"
  | "button_select"
  | "button_start"
  | "button_l"
  | "button_r"
  | "button_zl"
  | "button_zr"
  | "circle_pad"
  | "c_stick"
  | "button_debug"
  | "button_gpio14"
  | "button_home"
  | "button_power";